requirements.txt
accelerate==1.2.1
asposestorage==1.0.2
basicsr==1.4.2
diffusers==0.25.0
einops==0.8.1
lpips==0.1.4
matplotlib==3.7.3
numpy==1.24.4
openai_clip==1.0.1
opencv==4.7.0
opencv_python==4.8.0.74
Pillow==9.5.0
Pillow==12.0.0
ptflops==0.7.5
pyiqa==0.1.10
scipy==1.16.3
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.28.1
